more stuff

+++ Force/Blob.pde
@@ -27,7 +27,7 @@ class Blob extends Sprite
   }
   
   // Create a blob with a particular velocity vector.
-  Blob(int x, int y, int r, float dx, float dy, color c)
+  Blob(float x, float y, int r, float dx, float dy, color c)
   {
     super(x, y, dx, dy);
     myColor = c;
